fix: defmt 版本兼容 + board/pub API 重构 + 编译通过
- 固定 defmt =0.3.10 解决符号链接问题
- 添加 _defmt_timestamp 空实现 (省空间)
- board.rs: gpio_set/gpio_clr/spi0_transfer/uart API 公开化
- board.rs: write_reg 指针类型修正 (*const -> *mut)
- att7022e.rs: 扩展 +133 行
- hal.rs/display.rs/quectel.rs: 小修
- Release 编译: 3.5KB ARM binary

+++ firmware/src/display.rs
@@ -312,7 +312,7 @@ impl LcdPanel {
 
         // 写入段码
         for i in start..8 {
-            let dp = if dp_pos > 0 && (7 - i) == dp_pos { true } else { false };
+            let dp = if dp_pos > 0 && (7 - i) == dp_pos as usize { true } else { false };
             self.write_digit(start_digit + (i - start) as u8, digits[i], dp);
         }
     }
